♻️ Refactor: 캐러셀 커스텀 훅 분리
